+++ include/hm_queue.h
@@ -45,8 +45,8 @@ typedef struct hm_queue {
  * Simple functions
  */
 
-#define hm_queue_size(s) ((s)->size)
-#define hm_queue_capacity(s) ((s)->capacity)
+extern size_t hm_queue_size(hm_queue* queue);
+extern size_t hm_queue_capacity(hm_queue* queue);