+++ include/hm_base.h
@@ -12,9 +12,9 @@
 #define HM_BASE_H
 
 #define HM_FOCX_VERSION_MAJOR 0
-#define HM_FOCX_VERSION_MINOR 9
+#define HM_FOCX_VERSION_MINOR 10
 #define HM_FOCX_VERSION_PATCH 0
-#define HM_FOCX_VERSION "0.9.0"
+#define HM_FOCX_VERSION "0.10.0"
 
 #include <stdio.h>